% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2_Random_Forest.R
\name{binarize_categorical}
\alias{binarize_categorical}
\title{Binarizes dataframe columns which are characters and have only two unique
values inside of themselves}
\usage{
binarize_categorical(data)
}
\arguments{
\item{data}{data.frame to binarize}
}
\value{
data.frame with binarized proper columns
}
\description{
Binarizes dataframe columns which are characters and have only two unique
values inside of themselves
}
\examples{
data <- read.csv("./archive/lisbon-houses.csv")
df   <- to_DataFrame(data, "Price")
scaled   <- scale_continuous(df, c(5, 6, 7, 8))
binarized <- binarize_categorical(scaled)
head(binarized)
}
