% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2_Random_Forest.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{Performs one hot encoding on given dataframe. OHE is performed on character
columns with more than two unique categories only.}
\usage{
one_hot_encode(data, target)
}
\arguments{
\item{data}{data.frame to binarize}

\item{target}{string providing target column name}
}
\value{
data.frame with one hot encoded categorical columns
}
\description{
Performs one hot encoding on given dataframe. OHE is performed on character
columns with more than two unique categories only.
}
\examples{
data <- read.csv("./archive/lisbon-houses.csv")
df   <- to_DataFrame(data, "Price")
scaled1   <- scale_continuous(df, c("Bedrooms", "Bathrooms", "AreaNet", "AreaGross"))
binarized <- binarize_categorical(scaled1)
encoded   <- one_hot_encode(binarized, target = "Price")
names(encoded)
head(encoded)
}
