% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2_Random_Forest.R
\name{preprocessing}
\alias{preprocessing}
\title{Performs all preprocessing methods on the given data. The preprocessing
methods are: conversion to DataFrame, managing null values, scaling continuous
data, binarizing categorical columns with 2 categories, one hot encoding other
columns.}
\usage{
preprocessing(
  data,
  target,
  to_scale = NULL,
  missing_num = FALSE,
  missing_cat = FALSE
)
}
\arguments{
\item{data}{data provided as list, matrix, data.table or data.frame}

\item{target}{string, name of the target column}

\item{to_scale}{vector of column indexes or column names to scale}

\item{missing_num}{additional parameter, value which describes missing value
numeric observations (other than NA)}

\item{missing_cat}{additional parameter, value which describes missing value
for character observations (other than NA)}
}
\value{
preprocessed dataset ready to split.
}
\description{
Performs all preprocessing methods on the given data. The preprocessing
methods are: conversion to DataFrame, managing null values, scaling continuous
data, binarizing categorical columns with 2 categories, one hot encoding other
columns.
}
\examples{
data     <- read.csv("./archive/lisbon-houses.csv")
data     <- preprocessing(data, "Price", to_scale = c("Bedrooms", "Bathrooms", "AreaNet", "AreaGross"))
data
}
