% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2_Random_Forest.R
\name{scale_continuous}
\alias{scale_continuous}
\title{Scales and standardizes the columns provided as a vector of column indexes
or column names.}
\usage{
scale_continuous(data, to_scale)
}
\arguments{
\item{data}{data.frame to scale}

\item{to_scale}{vector of column indexes or column names to scale}
}
\value{
data.frame with scaled columns
}
\description{
Scales and standardizes the columns provided as a vector of column indexes
or column names.
}
\examples{
data    <- read.csv("./archive/lisbon-houses.csv")
df      <- to_DataFrame(data, "Price")
df      <- null_exterminator(df_null, "Price")
scaled0 <- scale_continuous(df, c("District", "Bedrooms", "Bathrooms", "AreaNet", "AreaGross"))
scaled1 <- scale_continuous(df, c("Bedrooms", "Bathrooms", "AreaNet", "AreaGross"))
scaled2 <- scale_continuous(df, c(5, 6, 7, 8))
head(scaled0)
head(scaled1)
head(scaled2)
}
