% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2_Random_Forest.R
\name{to_DataFrame}
\alias{to_DataFrame}
\title{Automatically converts other types of data to the data.frame type.}
\usage{
to_DataFrame(data, target)
}
\arguments{
\item{data}{data provided as list, matrix, data.table or data.frame}

\item{target}{string, name of the target column}
}
\value{
returns data.frame object
}
\description{
Automatically converts other types of data to the data.frame type.
}
\examples{
data <- read.csv("./archive/lisbon-houses.csv")
df   <- to_DataFrame(data, "Price")

}
