% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM2_Random_Forest.R
\name{train_test_split}
\alias{train_test_split}
\title{Divides dataset into 4 subsets of two training sets and two
testing sets for x and y. The user can set train_set size}
\usage{
train_test_split(data, target, train_size = 0.75)
}
\arguments{
\item{data}{data.frame to split}

\item{target}{string, name of the target column}

\item{train_size}{double / float, value from range [0,1] which sets the
size of the training set}
}
\value{
list of: train_x, train_y, test_x, test_y data.frames
}
\description{
Divides dataset into 4 subsets of two training sets and two
testing sets for x and y. The user can set train_set size
}
\examples{
data     <- read.csv("./archive/lisbon-houses.csv")
data     <- preprocessing(data, "Price", to_scale = c("Bedrooms", "Bathrooms", "AreaNet", "AreaGross"))
data     <- train_test_split(data = data, target = "Price",  train_size = 0.7)
data
}
